0.3.1_SU - Finalized formatting options, formatted Main
"Why did you choose formatting defaults instead of using rustfmt's
defaults???"
Because rustfmt's an absolute joke of a formatting tool with silly
defaults, the current styling standard is up in the air and not
concrete or consistent at all, and it will ensure that there's at least
some modicum of standard formatting for this project, at least. When the
only other possibility of formatting is to use the absolute trash known
as JetBrains' IDEs, or god forbid the RustRover abomination, then you
know that standards need to be declared.