0.3.2 - Init, added starting library, refactoring.
This update introduces the local library for GAT, including a very rough (and alpha) quantified layout for GWM's reponses via IPC. The objects are mainly not in use yet, but what few working parts I made I've implemented in the main crate. Not really worth releasing a binary for it yet, but the gist is to make something of a functional library for GWM's IPC. However, I plan on refactoring/remaking the entire listed approach, soooooooo yeah.