0.3.2 - Full anyhow integration, refactoring
Well, I got tired of manually managing the different error types, so I
decided to just convert everything over to `anyhow`. Not the greatest
for performance, or so I hear, but at least there will be some degree of
consistency. Besides, the only other alternative is to either find or
make a reduction step for the different errors, converting them to
`std::result::Result`s, and that's not worth it for this project.

Also, I've disregarded the liberal usage of `if let`, since almost every
instance that I've tried now requires an `else` clause, when most didn't
before. No idea if it's the result of a Rust change, since I've updated
multiple times since the start of this project, but it's possible.

Still having a think about the library structure and design, I _really_
hate the idea of doing it with this nested object/C#-ish method. I think
I might have found a solution, however.